requirements.txt
segno>=1.6.1
Pillow>=10.0.0
cairosvg>=2.7.0